+++ cli/src/from/arcs.rs
@@ -199,11 +199,8 @@ pub fn from_csv(global_args: GlobalArgs, args: CliArgs, file: impl BufRead) -> R
     }
 
     thread_pool.install(|| {
-        builder.par_comp_lenders_endianness(
-            &g,
-            num_nodes,
-            &target_endianness.unwrap_or_else(|| BE::NAME.into()),
-        )
+        builder
+            .par_comp_lenders_endianness(&g, &target_endianness.unwrap_or_else(|| BE::NAME.into()))
     })?;
 
     // save the nodes