+++ cli/src/to/bvgraph.rs
@@ -96,7 +96,6 @@ where
                 );
                 builder.par_comp_lenders_endianness(
                     &sorted,
-                    sorted.num_nodes(),
                     &target_endianness.unwrap_or_else(|| BE::NAME.into()),
                 )
             })?;
@@ -104,7 +103,6 @@ where
             thread_pool.install(|| {
                 builder.par_comp_lenders_endianness(
                     &graph,
-                    graph.num_nodes(),
                     &target_endianness.unwrap_or_else(|| BE::NAME.into()),
                 )
             })?;
@@ -131,7 +129,6 @@ where
             thread_pool.install(|| {
                 builder.par_comp_lenders_endianness(
                     &permuted,
-                    permuted.num_nodes(),
                     &target_endianness.unwrap_or_else(|| BE::NAME.into()),
                 )
             })?;
@@ -139,7 +136,6 @@ where
             thread_pool.install(|| {
                 builder.par_comp_lenders_endianness(
                     &seq_graph,
-                    seq_graph.num_nodes(),
                     &target_endianness.unwrap_or_else(|| BE::NAME.into()),
                 )
             })?;