+++ cli/src/transform/transpose.rs
@@ -89,7 +89,6 @@ where
     thread_pool.install(|| {
         builder.par_comp_lenders_endianness(
             &sorted,
-            sorted.num_nodes(),
             &target_endianness.unwrap_or_else(|| BE::NAME.into()),
         )
     })?;