+++ webgraph/src/utils/par_sort_pairs.rs
@@ -161,13 +161,14 @@ impl ParSortPairs {
     /// [`expected_num_pairs`](ParSortPairs::expected_num_pairs) can be used to
     /// customize the instance.
     ///
-    /// This method will return an error if the number of CPUs
-    /// returned by [`num_cpus::get()`](num_cpus::get()) is zero.
+    /// This method will return an error if [`rayon::current_num_threads`]
+    /// returns zero.
     pub fn new(num_nodes: usize) -> Result<Self> {
         Ok(Self {
             num_nodes,
             expected_num_pairs: None,
-            num_partitions: NonZeroUsize::new(num_cpus::get()).context("zero CPUs")?,
+            num_partitions: NonZeroUsize::new(rayon::current_num_threads())
+                .context("No Rayon threads")?,
             memory_usage: MemoryUsage::default(),
         })
     }
@@ -186,7 +187,7 @@ impl ParSortPairs {
     ///
     /// This is the number of iterators in the resulting [`SplitIters`].
     ///
-    /// Defaults to `num_cpus::get()`.
+    /// Defaults to [`rayon::current_num_threads`].
     pub fn num_partitions(self, num_partitions: NonZeroUsize) -> Self {
         Self {
             num_partitions,